+++ src/main/java/com/vsquad/iroas/aggregate/entity/Prop.java
@@ -22,19 +22,19 @@ public class Prop {
 
     @Column(name = "PROP_LOCATION_X")
     @Comment("PROP 위치 값 x")
-    private Float propLocationX;
+    private Double propLocationX;
 
     @Column(name = "PROP_LOCATION_Y")
     @Comment("PROP 위치 값 y")
-    private Float propLocationY;
+    private Double propLocationY;
 
     @Column(name = "PROP_LOCATION_Z")
     @Comment("PROP 위치 값 z")
-    private Float propLocationZ;
+    private Double propLocationZ;
 
     @Column(name = "PROP_YAW_VALUE")
     @Comment("Z축을 기준 회전 값")
-    private Float propYawValue;
+    private Double propYawValue;
 
     @Column(name = "CREATOR_MAP_ID")
     @Comment("크리에이터 툴에 의해 생성된 맵 식별자")
